\name{detMatrix}
\Rdversion{1.1}
\alias{detMatrix}
\title{
  calcualte the determinant of a square matrix
}
\description{
  Use LU decomposition to calculate the determinant of a square matrix
}
\usage{
detMatrix(X)
}
\arguments{
  \item{X}{ a square matrix }
}
\value{
  a real number of the determinant of the matrix
}
\author{ Wei Sun }
\keyword{ utilities }
