\name{mirror}
\alias{mirror}
\title{ mirror }
\description{
  Mirror the matrix of a graph about its diagonal 
}
\usage{
mirror(graph)
}
\arguments{
  \item{graph}{ The matrix of a graph }
}

\value{
\item{graph}{ Transformed matrix of the graph.}
}

\author{ Wei Sun <wsun@bios.unc.edu> }
\keyword{ methods }
