\name{print.MAL}
\alias{print.MAL}
\title{
  print an MAL object
}
\description{
  Print out all the elements of an MAL object, except 'yt', which is usually a big matrix. 
}
\usage{
  \method{print}{MAL}(x, ...)
}
\arguments{
  \item{x}{ an MAL object. }
  \item{...}{ further arguments passed to or from other methods. }
}
\author{ Wei Sun }
\keyword{ utilities }
