\name{print.mGAP}
\alias{print.mGAP}
\title{
  print an mGAP object
}
\description{
  Print out all the elements of an mGAP object, except 'yt', which is usually a big matrix. 
}
\usage{
  \method{print}{mGAP}(x, ...)
}
\arguments{
  \item{x}{ an mGAP object. }
  \item{...}{ further arguments passed to or from other methods. }
}
\author{ Wei Sun }
\keyword{ utilities }
