\name{print.pSDR}
\alias{print.pSDR}
\title{
  print an pSDR object
}
\description{
  Print out all the elements of an pSDR object, except 'yt', which is usually a big matrix. 
}
\usage{
  \method{print}{pSDR}(x, ...)
}
\arguments{
  \item{x}{ an pSDR object. }
  \item{...}{ further arguments passed to or from other methods. }
}
\author{ Wei Sun }
\keyword{ utilities }
